+++ contracts/interfaces/IStarter.sol
@@ -4,5 +4,5 @@ pragma solidity >=0.4.22 <0.9.0;
 interface IStarter {
     function startGame(uint256 card) external payable;
 
-    function cancelGame(uint256 gameId) external;
+    function cancelGame(uint64 gameId) external;
 }
